pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "infra-x"
version = "0.1.0"
description = "Open-source AI Terraform generator with versioned blueprints. Bring your own LLM (Ollama / Anthropic / OpenAI). Local-first."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "infra-x contributors" }]
keywords = [
    "terraform",
    "infrastructure-as-code",
    "iac",
    "hcl",
    "ai",
    "llm",
    "ollama",
    "anthropic",
    "openai",
    "aws",
    "gcp",
    "blueprints",
    "code-generation",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
    "jinja2>=3.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "build>=1.2.0",
    "twine>=5.1.0",
]

[project.scripts]
infra-x = "infra_x.cli:app"

[project.urls]
Homepage = "https://github.com/infra-x/infra-x"
Documentation = "https://github.com/infra-x/infra-x#readme"
Repository = "https://github.com/infra-x/infra-x"
Issues = "https://github.com/infra-x/infra-x/issues"
Changelog = "https://github.com/infra-x/infra-x/blob/main/CHANGELOG.md"

# --- Hatchling build config -------------------------------------------------
# `packages = ["infra_x"]` ships the entire package tree, including the YAML
# blueprint catalog under `infra_x/blueprints/catalog/`. No extra config needed
# — hatchling's wheel builder picks up non-Python files inside the package.

[tool.hatch.build.targets.wheel]
packages = ["infra_x"]

[tool.hatch.build.targets.sdist]
include = [
    "/infra_x",
    "/tests",
    "/README.md",
    "/LICENSE",
    "/Makefile",
    "/pyproject.toml",
]

# --- Lint / format ----------------------------------------------------------

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
